+++ src/core/http/http.py
@@ -17,7 +17,7 @@
 """
 
 from urllib3 import HTTPConnectionPool, PoolManager, Timeout
-from urllib3.exceptions import MaxRetryError, ReadTimeoutError, ConnectTimeoutError, HostChangedError
+from urllib3.exceptions import DecodeError, MaxRetryError, ReadTimeoutError, ConnectTimeoutError, HostChangedError
 from src.core import helper
 from .exceptions import HttpRequestError
 from .providers import DebugProvider
@@ -161,6 +161,9 @@ class HttpRequest(RequestProvider, DebugProvider):
 
         except ReadTimeoutError:
             self.__tpl.warning(key='read_timeout_error', url=url)
+
+        except DecodeError:
+            self.__tpl.warning(key='decode_error', url=url)
 
         except ConnectTimeoutError:
             self.__tpl.warning(key='connection_timeout_error', url=url)